Moved internal functions from inside the instance resource to the resource package to be reused when I'll add the master resource
