+++ service/controller/resource/instance/create_cordon_old_workers.go
@@ -17,7 +17,7 @@ import (
 	"github.com/giantswarm/azure-operator/pkg/project"
 	"github.com/giantswarm/azure-operator/service/controller/controllercontext"
 	"github.com/giantswarm/azure-operator/service/controller/key"
-	"github.com/giantswarm/azure-operator/service/controller/resource/instance/internal/state"
+	"github.com/giantswarm/azure-operator/service/controller/resource/internal/state"
 )
 
 const (
