+++ service/controller/resource/instance/create_provisioning_successful.go
@@ -3,7 +3,7 @@ package instance
 import (
 	"context"
 
-	"github.com/giantswarm/azure-operator/service/controller/resource/instance/internal/state"
+	"github.com/giantswarm/azure-operator/service/controller/resource/internal/state"
 )
 
 func (r *Resource) provisioningSuccessfulTransition(ctx context.Context, obj interface{}, currentState state.State) (state.State, error) {
