+++ service/controller/resource/internal/state/funcs.go
