+++ service/controller/resource/internal/state/funcs_test.go
